Add check what read after set readed byte in 0x4a 0x00.
After tests revealed that, when requesting the version number via the
entry of the requested bit in register 0x4a 0x00, does not always return
the correct result (all char of version) but all return codes are
good. The only thing that is stable when the error is the result is
equal to 0x00, rather than the expected symbol.
Know chip version is desirable, because not only in our devices are used
2 (or more) different, but module saa7115 makes the verification of the above
algorithm and this method, in theory, should work in that code.
If 0x00 is returned i check: timeouts, rereadin, or write twice in registry -
situation not change. So was added check for proper reading registry while writing to
these registers: 0x4a 0x00, on failure - repeat write - most stable
result now.

Manouchehri/smi2021#15
This hack - if someone is good - always welcome :)